Add announcment_posts table in migrations

+++ backend/db/migrations/00003_create_posts.sql
@@ -2,7 +2,7 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS posts (
     id SERIAL PRIMARY KEY,
-    author_id INT64 NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    author_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     type TEXT NOT NULL CHECK (type IN ('announcement', 'report','marketplace','openjio')),
     title TEXT NOT NULL,
     body TEXT NOT NULL,
