+++ backend/db/sqlc/queries.sql.go
@@ -30,7 +30,7 @@ const getUserByID = `-- name: GetUserByID :one
 SELECT id, name, email, role FROM users WHERE id = $1
 `
 
-func (q *Queries) GetUserByID(ctx context.Context, id int64) (User, error) {
+func (q *Queries) GetUserByID(ctx context.Context, id int32) (User, error) {
 	row := q.db.QueryRow(ctx, getUserByID, id)
 	var i User
 	err := row.Scan(
